♻️ refactor: article関連の関数を整理
