--- src/utils/article/index.ts
@@ -1,1 +1,0 @@
-export { isDraft } from "./isDraft";
